+++ src/test_Clipping_MISR_MODIS.cpp
@@ -29,13 +29,13 @@ int main(int argc, char ** argv){
 		exit(1);
 	}
 	
-	int nCellsrc;
+	int64_t nCellsrc;
 	double* src_lat;
 	double* src_long;
 	src_lat = get_misr_lat(src_file, "L", &nCellsrc);
 	src_long = get_misr_long(src_file, "L", &nCellsrc);
 	
-	int nCelldest;
+	int64_t nCelldest;
 	double* dest_lat;
 	double* dest_long;
 	
@@ -52,13 +52,13 @@ int main(int argc, char ** argv){
 	}
 
 	double* projected_Rad_Out;
-	int * tarNNSouID;
+	int64_t * tarNNSouID;
 	//double** p_src_lat = &src_lat;
 	//double** p_src_lon = &src_long;
 	double** p_src_lat = &src_lat;
 	double** p_src_lon = &src_long;
 	
-	tarNNSouID = (int *)malloc(sizeof(int) * nCelldest);
+	tarNNSouID = (int64_t *)malloc(sizeof(int64_t) * nCelldest);
 	
 	printf("nearest neighbor\n");
 	//nearestNeighbor(p_src_lat, p_src_lon, nCellsrc, dest_lat, dest_long, tarNNSouID, nCelldest, 1000);
@@ -78,13 +78,13 @@ int main(int argc, char ** argv){
 	
 	src_rad = get_misr_rad(src_file, "AN", "L", "Blue_Radiance", &nCellsrc);
 	
-	int nCelldest_rad;
+	int64_t nCelldest_rad;
 	double* dest_rad;
 	std::vector<std::string> bands = {"8"};
 	dest_rad = get_modis_rad(src_file, "_1KM", bands, bands.size(), &nCelldest_rad);
 	
 	double* src_rad_out = (double *)malloc(sizeof(double) * nCelldest);
-	int new_misr_size = nCelldest;
+	int64_t new_misr_size = nCelldest;
 	//Interpolating
 //	int * nsrcPixels;
 	printf("interpolating\n");
